Renamed aws param store file
